requirements.txt
Django==4.2.4
psycopg2==2.9.7
Pillow==10.0.0
django-multiselectfield==0.1.12